% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{fromTicFormat}
\alias{fromTicFormat}
\title{Create a date from a Tic compatable string}
\usage{
fromTicFormat(date)
}
\arguments{
\item{date}{- date string to convert}
}
\value{
The parsed date
}
\description{
Create a date from a Tic compatable string
}
\examples{
fromTicFormat("2019.01.01")
}
