% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{getAdjacentWeekdays}
\alias{getAdjacentWeekdays}
\title{Return a list of the n adjacent weekdays}
\usage{
getAdjacentWeekdays(date, n, forward = TRUE)
}
\arguments{
\item{date}{- the date to offset}

\item{n}{- the number of days to offset (a negative n means reverse the meaning of the forward parameter)}

\item{forward}{- forward or backward offset}
}
\value{
The next weekday(s)
}
\description{
Return a list of the n adjacent weekdays
}
\examples{
getAdjacentWeekdays(today(), 4, forward = FALSE)
getAdjacentWeekdays(today(), -4)
}
