% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{getNextWeekday}
\alias{getNextWeekday}
\title{Get the next weekday}
\usage{
getNextWeekday(date)
}
\arguments{
\item{date}{- the date to offset}
}
\value{
The next weekday
}
\description{
Get the next weekday
}
\examples{
getNextWeekday(today())
}
