% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{getPreviousWeekday}
\alias{getPreviousWeekday}
\title{Get the previous weekday}
\usage{
getPreviousWeekday(date)
}
\arguments{
\item{date}{- the date to offset}
}
\value{
The previous weekday
}
\description{
Get the previous weekday
}
\examples{
getPreviousWeekday(today())
}
