% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{getWeekdays}
\alias{getWeekdays}
\title{Get all the week days between two points}
\usage{
getWeekdays(startDate, endDate)
}
\arguments{
\item{startDate}{- start date}

\item{endDate}{- end date}
}
\value{
A list of Dates
}
\description{
Get all the week days between two points
}
\examples{
getWeekdays("2019.01.01", "2019.01.05")
}
