% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{isWeekend}
\alias{isWeekend}
\title{Check if a date falls on a weekend}
\usage{
isWeekend(date)
}
\arguments{
\item{date}{- the date(s) to test}
}
\value{
A boolean (or list of booleans)
}
\description{
Check if a date falls on a weekend
}
\examples{
isWeekend(c(ymd("20190505"), ymd("20190506")))
}
