% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{offsetByWeekdays}
\alias{offsetByWeekdays}
\title{Return the nth weekday from now}
\usage{
offsetByWeekdays(date, n, forward = TRUE)
}
\arguments{
\item{date}{- the date to offset}

\item{n}{- the number of days to offset (a negative n means reverse the meaning of the forward parameter)}

\item{forward}{- forward or backward offset}
}
\value{
The resultant weekday
}
\description{
Return the nth weekday from now
}
\examples{
offsetByWeekdays(today(), 4, forward = FALSE)
offsetByWeekdays(today(), -4)
}
