% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateTime.R
\name{toTicFormat}
\alias{toTicFormat}
\title{Convert a date into a Tic compatable string}
\usage{
toTicFormat(date)
}
\arguments{
\item{date}{- date to convert}
}
\value{
The string to pass to tic
}
\description{
Convert a date into a Tic compatable string
}
\examples{
toTicFormat(ymd("2019-01-01"))
}
